requirements.txt
python-bitcoinrpc
matplotlib
